requirements.txt
fastapi
uvicorn
google-genai
dotenv
databricks #maybe not neccessary who knows atp 
databricks.sql